+++ src/frontend/A64/translate/impl/simd_two_register_misc.cpp
@@ -384,7 +384,7 @@ bool TranslatorVisitor::FNEG_2(bool Q, bool sz, Vec Vn, Vec Vd) {
 
     const size_t datasize = Q ? 128 : 64;
     const size_t esize = sz ? 64 : 32;
-    const size_t mask_value = esize == 64 ? 0x8000000000000000 : 0x8000000080000000;
+    const size_t mask_value = static_cast<size_t>(esize == 64 ? 0x8000000000000000 : 0x8000000080000000);
 
     const IR::U128 operand = V(datasize, Vn);
     const IR::U128 mask = ir.VectorBroadcast(esize, I(esize, mask_value));
